checking in version of code with mostly working repeater plugin- need to add a handler for checking for multiple responses to same message
