+++ .gitignore
@@ -3,3 +3,4 @@ db/
 distributed/manager/*.class
 distributed/manager/debug
 distributed/plugins/*.class
+distributed/debug
