+++ distributed/plugins/QueueServiceThread.java
@@ -1,61 +1,66 @@
 
 
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.ArrayList;
 import java.util.List;
 
-   class QueueServiceThread extends Thread 
-   { 
-      BlockingQueue<VdmsTransaction> queue;
-      TestPlugin manager;
-      int matchType;
-
-
-      public QueueServiceThread(BlockingQueue<VdmsTransaction> nQueue, TestPlugin nManager, int nMatchType)
+class QueueServiceThread extends Thread 
+{ 
+   BlockingQueue<VdmsTransaction> queue;
+   TestPlugin manager;
+   int matchType;
+   
+   
+   public QueueServiceThread(BlockingQueue<VdmsTransaction> nQueue, TestPlugin nManager, int nMatchType)
+   {
+      queue = nQueue;
+      manager = nManager;
+      matchType = nMatchType;
+   }
+   
+   public void run()
+   {
+      VdmsTransaction message;
+      List<PublisherServiceThread> publishList;
+      List<SubscriberServiceThread> subscribeList;
+      
+      try
       {
-         queue = nQueue;
-         manager = nManager;
-         matchType = nMatchType;
-      }
-
-      public void run()
-      {
-         VdmsTransaction message;
-         List<ServerServiceThread> publishList;
-
-         try
+         while(true)
          {
-            while(true)
+            message = queue.take();
+            
+            //Get any new publishers that may exist
+            if(matchType == 0)
             {
-               message = queue.take();
-
-               //Get any new publishers that may exist
-               if(matchType == 0)
+               subscribeList = manager.GetConsumerList();
+               //Publish to all of the associated threads
+               for(int i = 0; i < subscribeList.size(); i++)
                {
-                  publishList = manager.GetConsumerList(); 
+                  subscribeList.get(i).Publish(message);
                }
-               else
-               {
-                  //Before sending data, get a list of any potential new Producers
-                  publishList = manager.GetProducerList();
-               }
-
+               
+            }
+            else
+            {
+               //Before sending data, get a list of any potential new Producers
+               publishList = manager.GetProducerList();
                //Publish to all of the associated threads
                for(int i = 0; i < publishList.size(); i++)
                {
                   publishList.get(i).Publish(message);
                }
             }
-
-         }
-         catch(InterruptedException e)
-         {
-            this.interrupt();
             
          }
-
-
+         
       }
-
+      catch(InterruptedException e)
+      {
+         this.interrupt();
+         
+      }
+      
+      
    }
+   
+}
