+++ distributed/plugins/VdmsTransaction.java
@@ -1,6 +1,3 @@
-import java.util.ArrayList;
-import java.util.List;
-
 class VdmsTransaction
 {
     byte size[];
